+++ src/imxrt1010/lpi2c.rs
@@ -12,32 +12,32 @@ use crate::{
 //
 
 // SCL
-i2c!(module: U1, alt: 0, pad: GPIO_AD_14,    signal: Scl, daisy: DAISY_LPI2C1_SCL_GPIO_AD_14);
-i2c!(module: U1, alt: 1, pad: GPIO_SD_06,    signal: Scl, daisy: DAISY_LPI2C1_SCL_GPIO_SD_06);
-i2c!(module: U1, alt: 1, pad: GPIO_12,       signal: Scl, daisy: DAISY_LPI2C1_SCL_GPIO_12);
-i2c!(module: U1, alt: 3, pad: GPIO_02,       signal: Scl, daisy: DAISY_LPI2C1_SCL_GPIO_02);
+i2c!(module: U1, alt: 0, pad: GPIO_AD_14,    signal: Scl, daisy: Some(DAISY_LPI2C1_SCL_GPIO_AD_14));
+i2c!(module: U1, alt: 1, pad: GPIO_SD_06,    signal: Scl, daisy: Some(DAISY_LPI2C1_SCL_GPIO_SD_06));
+i2c!(module: U1, alt: 1, pad: GPIO_12,       signal: Scl, daisy: Some(DAISY_LPI2C1_SCL_GPIO_12));
+i2c!(module: U1, alt: 3, pad: GPIO_02,       signal: Scl, daisy: Some(DAISY_LPI2C1_SCL_GPIO_02));
 
 // SDA
-i2c!(module: U1, alt: 0, pad: GPIO_AD_13,    signal: Sda, daisy: DAISY_LPI2C1_SDA_GPIO_AD_13);
-i2c!(module: U1, alt: 1, pad: GPIO_SD_05,    signal: Sda, daisy: DAISY_LPI2C1_SDA_GPIO_SD_05);
-i2c!(module: U1, alt: 1, pad: GPIO_11,       signal: Sda, daisy: DAISY_LPI2C1_SDA_GPIO_11);
-i2c!(module: U1, alt: 3, pad: GPIO_01,       signal: Sda, daisy: DAISY_LPI2C1_SDA_GPIO_01);
+i2c!(module: U1, alt: 0, pad: GPIO_AD_13,    signal: Sda, daisy: Some(DAISY_LPI2C1_SDA_GPIO_AD_13));
+i2c!(module: U1, alt: 1, pad: GPIO_SD_05,    signal: Sda, daisy: Some(DAISY_LPI2C1_SDA_GPIO_SD_05));
+i2c!(module: U1, alt: 1, pad: GPIO_11,       signal: Sda, daisy: Some(DAISY_LPI2C1_SDA_GPIO_11));
+i2c!(module: U1, alt: 3, pad: GPIO_01,       signal: Sda, daisy: Some(DAISY_LPI2C1_SDA_GPIO_01));
 
 //
 // I2C2
 //
 
 // SCL
-i2c!(module: U2, alt: 0, pad: GPIO_AD_08,    signal: Scl, daisy: DAISY_LPI2C2_SCL_GPIO_AD_08);
-i2c!(module: U2, alt: 1, pad: GPIO_SD_08,    signal: Scl, daisy: DAISY_LPI2C2_SCL_GPIO_SD_08);
-i2c!(module: U2, alt: 3, pad: GPIO_AD_02,    signal: Scl, daisy: DAISY_LPI2C2_SCL_GPIO_AD_02);
-i2c!(module: U2, alt: 3, pad: GPIO_10,       signal: Scl, daisy: DAISY_LPI2C2_SCL_GPIO_10);
+i2c!(module: U2, alt: 0, pad: GPIO_AD_08,    signal: Scl, daisy: Some(DAISY_LPI2C2_SCL_GPIO_AD_08));
+i2c!(module: U2, alt: 1, pad: GPIO_SD_08,    signal: Scl, daisy: Some(DAISY_LPI2C2_SCL_GPIO_SD_08));
+i2c!(module: U2, alt: 3, pad: GPIO_AD_02,    signal: Scl, daisy: Some(DAISY_LPI2C2_SCL_GPIO_AD_02));
+i2c!(module: U2, alt: 3, pad: GPIO_10,       signal: Scl, daisy: Some(DAISY_LPI2C2_SCL_GPIO_10));
 
 // SDA
-i2c!(module: U2, alt: 0, pad: GPIO_AD_07,    signal: Sda, daisy: DAISY_LPI2C2_SDA_GPIO_AD_07);
-i2c!(module: U2, alt: 1, pad: GPIO_SD_07,    signal: Sda, daisy: DAISY_LPI2C2_SDA_GPIO_SD_07);
-i2c!(module: U2, alt: 3, pad: GPIO_AD_01,    signal: Sda, daisy: DAISY_LPI2C2_SDA_GPIO_AD_01);
-i2c!(module: U2, alt: 3, pad: GPIO_09,       signal: Sda, daisy: DAISY_LPI2C2_SDA_GPIO_09);
+i2c!(module: U2, alt: 0, pad: GPIO_AD_07,    signal: Sda, daisy: Some(DAISY_LPI2C2_SDA_GPIO_AD_07));
+i2c!(module: U2, alt: 1, pad: GPIO_SD_07,    signal: Sda, daisy: Some(DAISY_LPI2C2_SDA_GPIO_SD_07));
+i2c!(module: U2, alt: 3, pad: GPIO_AD_01,    signal: Sda, daisy: Some(DAISY_LPI2C2_SDA_GPIO_AD_01));
+i2c!(module: U2, alt: 3, pad: GPIO_09,       signal: Sda, daisy: Some(DAISY_LPI2C2_SDA_GPIO_09));
 
 mod daisy {
     #![allow(unused)]